Extracted renderer code into a child of compiler
